integration tests: Splitup cassandra connection abstraction
